Extract cubic Bezier binomial intersection tests to a separate file
